setup.py: Fix Windows build dll path.
In the workflow it placed the library at:

D:\a\libusb-package\libusb-package\src\libusb\build\v143\Win32\Release\dll\libusb-1.0.dll

So the "build\v143\" part must be new when libusb rearranged how they
build stuff.  I'm not sure if "v143" is something we can rely on, so I
am now globbing with a * there.